test: convert tests to jasmine 3
