+++ .eslintrc.json
@@ -111,7 +111,6 @@
 
 		"arrow-spacing": 2,
 		"prefer-const": 1,
-		"prefer-destructuring": 1,
 		"prefer-rest-params": 1,
 		"prefer-spread": 1,
 		"prefer-template": 1,
