+++ spec/helper.js
@@ -10,53 +10,64 @@ A collection of methods for retrieving information about the user's system for
 bug report purposes.
 */
 
+async function jsonPromise(object) {
+	return {
+		ok: true,
+		async json() {
+			return object;
+		}
+	};
+}
+async function textPromise(text) {
+	return {
+		ok: true,
+		async text() {
+			return text;
+		}
+	};
+}
+
 module.exports = {
 
 	generateException() {
 		try {
-			return a + 1;
+			// eslint-disable-next-line no-undef, no-unused-expressions
+			a + 1;
 		} catch (e) {
 			const errMsg = `${e.toString()} in ${process.env.ATOM_HOME}/somewhere`;
-			return window.onerror.call(window, errMsg, "/dev/null", 2, 3, e);
+			window.onerror.call(window, errMsg, "/dev/null", 2, 3, e);
 		}
 	},
 
 	// shortenerResponse
 	// packageResponse
 	// issuesResponse
-	generateFakeFetchResponses(options) {
-		if (!window.fetch.isSpy) {
-			spyOn(window, "fetch"); 
+	generateFakeFetchResponses(options = {}) {
+		if (!window.fetch.and) {
+			spyOn(window, "fetch");
 		}
 
-		return fetch.andCallFake(function (url) {
+		fetch.and.callFake(function (url) {
 			if (url.indexOf("is.gd") > -1) {
-				return textPromise((options != null ? options.shortenerResponse : undefined) != null ? (options != null ? options.shortenerResponse : undefined) : "http://is.gd/cats");
+				return textPromise(options.shortenerResponse ? options.shortenerResponse : "http://is.gd/cats");
 			}
 
 			if (url.indexOf("atom.io/api/packages") > -1) {
-				return jsonPromise((options != null ? options.packageResponse : undefined) != null ? (options != null ? options.packageResponse : undefined) : {
+				return jsonPromise(options.packageResponse ? options.packageResponse : {
 					repository: {url: "https://github.com/UziTech/notifications"},
 					releases: {latest: "0.0.0"}
 				});
 			}
 
 			if (url.indexOf("atom.io/api/updates") > -1) {
-				return (jsonPromise((options != null ? options.atomResponse : undefined) != null ? (options != null ? options.atomResponse : undefined) : {name: atom.getVersion()}));
+				return (jsonPromise(options.atomResponse ? options.atomResponse : {name: atom.getVersion()}));
 			}
 
-			if ((options != null ? options.issuesErrorResponse : undefined) != null) {
-				return Promise.reject(options != null ? options.issuesErrorResponse : undefined);
+			if (options.issuesErrorResponse) {
+				return Promise.reject(options.issuesErrorResponse);
 			}
 
-			return jsonPromise((options != null ? options.issuesResponse : undefined) != null ? (options != null ? options.issuesResponse : undefined) : {items: []});
+			return jsonPromise(options.issuesResponse ? options.issuesResponse : {items: []});
 		});
 	}
 };
-
-var jsonPromise = object => Promise.resolve({ok: true, json() {
-	return Promise.resolve(object); 
-}});
-var textPromise = text => Promise.resolve({ok: true, text() {
-	return Promise.resolve(text); 
-}});
