+++ spec/runner.js
@@ -2,9 +2,16 @@ const {createRunner} = require("atom-jasmine3-test-runner");
 
 module.exports = createRunner({
 	specHelper: {
-		jasminePass: true,
 		customMatchers: true,
 		attachToDom: true,
 		ci: true
 	},
+}, () => {
+	beforeEach(() => {
+		jasmine.clock().install();
+	});
+
+	afterEach(() => {
+		jasmine.clock().uninstall();
+	});
 });
